CS655: lev_comp:Fix traps replacing stairs/ladders
create_trap() now checks to make sure it won't overwrite stairs or
ladders.

Compared to UnNetHack, the try count check and increment for the
checking loop is put before the stair and ladder checks, otherwise a
room filled with stairs/ladders will cause an infinite loop: the exact
case the try count is trying to avoid!

According to UnNetHack, this is from SporkHack r792.