CS678: Count identical messages for dumping
Last messages shown in the dumplog now compress repeated messages, like
this: "This door is locked. (12x)", so now it shows the last 30
different messages.

Original implementation to go with NitroHack's original dumplog
implementation.

This also fixes a bug in upstream NitroHack with Norep() not detecting
repeated messages properly; it was checking one-ahead of where it was
supposed to in the circular buffer of recent messages.

NitroHack doesn't seem to save or restore this recent messages buffer
either, which I'm sure is a bug, but I haven't done anything about that
here.