Make paranoid_trap smarter about prompting
paranoid_trap now suppresses its prompt in situations it wouldn't help.

The prompt no longer appears if the trap is a Vibrating Square or the
player character has flight/levitation and the trap is a squeaky board,
a bear trap, a spiked pit, a trapdoor, or (outside of Sokoban) a pit or
hole.

The prompt no longer appears if the player character is already trapped
and wouldn't move into the other trap's direction anyway.

The prompt no longer appears if a monster or invisible symbol is on the
trap.

The last two cases above are supposed to suppress the prompt for any
known non-move, but there are lots of corner cases and complicated
checks that aren't covered, so this just handles the most common cases.

This is meant to go with commit 000f36287c (CS625 add paranoid_trap
option).