CS704: Move paranoid_trap check after attack check
This is a better way to avoid the prompt if a monster is on the trap.

Compared to UnNetHack, the code here is moved after the u.utrap check
too, to avoid the prompt when currently stuck in a trap as well.  It
also rolls in the smarter paranoid_trap behavior I wrote and cleans up
its logic a bit, but skips that behavior if hallucinating (the whole
point of hallucination is to do things like that).