CS716: Hide random descriptions of Sokoban prizes
This replaces the replaces the random description of Sokoban prizes with
"sokoban amulet/bag/cloak" when used in messages, e.g. far-look.

I assume this is stop prevent players finding an object with an
identical description outside of Sokoban and thus identifying desirable
items without completing it.  If so, this puts graphical tiles UIs in an
awkward spot, since they'll still show the right randomized tile for
Sokoban prizes.