CS733: Stop uniques killing uniques w/ the Amulet
This stops e.g. luring Asmodeus to the Sanctum to fetch the Amulet for
you... not that that works in UnNetHack anyway, given that Cthulhu holds
it and is much stronger than the high priest of Moloch; in this case,
this change stops Asmodeus from even trying to kill Cthulhu.

The code originally in UnNetHack says:

    /* Don't let uniques attack other uniques with the amulet
     * as long the unique with amulet can't fight back. */

But as far as I can see, there's no code nearby that checks if the
Amulet-holding unique can fight back or not, so I left it out of the
comment in this commit.