Clear new mem_obj bits when forgetting map
Specifically, this clears the bits for a pile of multiple stacks of
objects, and whether the object is a Sokoban prize when a map is cleared
by amnesia.

This is meant to go with commit 662a119c70 (CS715 color Sokoban prizes
bright magenta) and commit 644c88b3e7 (CS377,513 multiple items on floor
in inverse video).