CS906: Autoexplore: prefer exploring items
Now autoexplore won't just zip past items when the player sees them and
decide to come back to them at the other side of the level.

This uses INT_MAX from limits.h, and I have no idea which C standard
either is meant to be from.  Compiles and runs fine here, so I hope it
won't matter.

Compared to UnNetHack, autoexploring is a boolean instead of an int
masquerading as a boolean.