Fix paranoid_trap prompt suppression when running
This broke when the paranoid_trap check was moved down, since there was
code in between that called nomul(0, NULL), which nuked all signs that
the player was ever running, leading to the prompt appearing when the
player ran into a trap using Shift+direction.

This is meant to go with commit 719c74a709 (CS704 move paranoid_trap
check after attack check).