CS759,760: Change syms and location types to enums
This was already partly done in NitroHack for dungeon background and
feature symbols for changes allowing dungeon features to be targetted,
but this converts defines into enums for most of the rest of rm.h.

I assume this change was made in UnNetHack to prepare for the addition
of dead trees, which is coming next.