CS764:Add spellorder option: changes spell letters
spellorder changes the order of letters used by spell menus to whatever
the player wants.  This also lets spells be set to any desired letter,
in a roundabout way.

It can be set to a string of letters and letter ranges; the ranges can
even run backwards, but both ends have to be the same case.

Compared to UnNetHack, this original implementation accepts letter
ranges (with ghetto parser!), partial input (putting unused letters
after the option input) and validates input in NitroHack's new options
framework (no number or symbol accelerators, which makes entries
unselectable anyway even in the latest version of UnNetHack).