Fix paid object on bill when angering another shkp
Without this fix, angering a shopkeeper while holding unpaid objects of
another shopkeeper will mark the objects as unpaid, causing "Paid object
on bill??" warnings when trying to pay for them and giving them to the
player for free.

Also removes useless null shkp check in setpaid(), since everything that
calls it assumes shkp is non-null, and if it isn't the case it should
break loudly.

This bug also exists in vanilla NetHack, sourced to a FIXME comment that
has been around since 2002.  Reported by PaRaD0xx in FreeNode's #nethack
IRC channel.