CS776-779,781,787,825: DYWYPISI patch (by Jafet)
This shows unknown item details in square brackets at the end of games,
e.g. n - an [uncursed +0] long sword named Excalibur [Excalibur].  The
original patch is here:

    http://bilious.alt.org/?365

The UnNetHack version of this patch uses program_state.gameover instead
of a static variable to track when this should kick in, which this port
to NitroHack also adopts.

Compared to UnNetHack:

 * Uses of mons[].mname go through mons_mname() instead to ensure that
   corpses, tins and eggs of dragons get the right names, as part of the
   randomized dragon appearances system unique to this porting effort.

 * monstermoves is replaced with moves, since the former no longer
   exists.

 * Tabs and spaces indentation is used for much of the code instead of
   pure tabs like the code around it, since the tabs-only indentation
   and wide lines hurt readability, but so does any attempt to break the
   lines up.

 * A small change is made to support NitroHack's beatitude (BUC)
   coloring, since this patch replaces the full inventory ID at the end
   of each game.