CS794: Fix the fix for C343-235
C343-235: "Casting spell of protection gives incorrect message if hero
is swallowed or in rock."

This corrects the check to use "stone" versus "air" when embedded in
stone to use the player's position, otherwise, most of the time
protection is cast the message will mention stone instead of air.

This is meant to go with commit adce07daf5 (fix C343-235).

This bug also exists in upstream NitroHack.