Add autodigdown option
When false (default), it prevents digging a pit/hole when autodig is
true and player presses '>' (downwards).

I added this because having autodig on and trying to go downstairs while
not on stairs often caused me to waste many turns trying to create a pit
I never asked for, and turning autodig on and off was also annoying.