Prevent pet displacing stopping travel/autoexplore
This stops the game from pausing if your pet happens to be on the space
that travel/autoexplore wants to take you, making them much less
annoying to use with a pet on the level.

It makes it harder to use travel to find pets, but the other run
functions are unchanged and will stop under all of the same situations
as before.

This is meant to go with commit 621c5ffedd (CS747,748,765,768,815
autoexplore from NetHack4).