CS818: Stop player vampire rising as other monster
For example, a player vampire killed by a mummy will not create a
'vampire mummy' in their bones, because they don't exist in the game.
In fact, this commit stops any ghost or monster from being created with
vampire bones.

This is meant to go with commit 8e01be9e0d (CS503,508,510 add vampire
player race).