Add mnemonic accelerators to items action menu
The action menu that comes up when selecting an item in the inventory
now has mnemonic key accelerators, e.g. 'w - Wield...'

The keys are chosen to avoid overlap, so some choices might seem odd
without explanation:

 D - mapped to dip, since #dip has no dedicated key by default

 R - mapped to rub, not remove

 T - mapped to remove as well as take off, since NitroHack allows
     accessories to be taken off with T

 W - mapped to put on as well as wear, since NitroHack allows
     accessories to to be worn with W

 S - mapped to #sacrifice, even though S is usually save and exit

Implementing this would normally require a network protocol change, but
Daniel Thaler had the foresight to transmit all of the fields of the
nh_cmd_desc structs that transport object commands from the back-end to
the front-end, so I just used the defkey field instead.

This is meant to go with commit 7f91acf6cd (add item actions menu) and
commit 59363d190e (nh_get_object_commands).