CS844: Pressing cmd key for direction repeats last
This allows things like 'ff' to fire or 'zzz' to zap a wand on letter
'z' in the last direction.  It's even possible to 'strafe' with this.

Compared to UnNetHack, this is once again original code for NitroHack's
split architecture.  Also, this implementation shows a hint at the
direction prompt if a directional command is repeated.