Start characters with weapon stacks quivered
If a character begins with a stack of any kind of weapon, it's a safe
bet that they are meant to be thrown or fired.

This is meant to go with commit 1d6a7e15b7 (start characters with stacks
of daggers quivered), fixing orcish rogues not starting with their
orcish daggers quivered.