Sort lists of inventory objects consistently
The sortloot patch in 'loot' mode is meant to sort everything except
objects in the player's inventory, but Daniel Thaler's adaption sorts
lists of inventory objects in prompts, but not the full inventory list,
resulting in inventory objects appearing in one order in the sidebar,
and in another in prompts.

The rationale behind doing this instead of sorting the inventory is that
the inventory can be rearranged in the order the player prefers, whereas
sorting it clobbers that.

This is meant to go with commit 1ae9b87285 (add loot sorting).

This bug also exists in upstream NitroHack.