Greatly improve travel in Sokoban
This makes two changes to travel in pre-mapped levels:

 1. Boulders are treated like walls by the travel system.

 2. Set the seen vectors of all level locations so travelling to
    pre-mapped but 'unseen' locations can make use of the map data.
    (The lack of this caused the travel system to fail in its
    deterministic 1st phase [BFS from target to player] and fall back on
    its heuristic 2nd phase [floodfill from player and choose closest
    player-reachable location].)

Together, travelling in Sokoban will no longer try to lead players
through boulders, but if there's a long way around, it will always be
found.  This makes Sokoban levels faster to solve and less typo-prone.

Beware, travelling onto an orthogonally-adjacent boulder is still
treated like a normal move which pushes it.  I can't decide if this
should be removed to prevent Sokoban typos, or kept in so that
point-and-click interfaces can interact with adjacent spaces, so I'm
leaving it alone for now.