CS585: Add name-HP bar to classic status lines
Modelled after the HitPointBar patch as seen on NAO:

    http://bilious.alt.org/?192

That is, its bar is only as wide as the text it contains, versus the HP
bar of NitroHack's three-line status mode which is of a fixed width.

With use_inverse enabled, this colors the whole name and rank and uses
inverse to show the bar.  With use_inverse off, only part of the name
and rank is colored.