Color changes to attributes green/red
Attributes, AC, experience, money and score are colored green when they
improve, or red when they degrade.  The colors remain until the turn
count changes, and the difference is calculated across multi-turn
actions.

Interestingly, stats that color green in replays going forward color red
when going backwards, and jumping directly to a turn colors the changes
for everything in between.