CS846,847: Give gnomes candles sometimes
1 in 4 gnomes will carry candles (under their usual generation rules, so
half the time they do it will be 1, otherwise it will be 1 to 7).  25%
of the time these will be wax candles, otherwise they will be tallow.

Code is from NetHack 4:

    https://gitorious.org/nitrohack/ais523/commit/9dc7a4c2f7cc7920d1edebf58007e1bdffdfd759

Compared to UnNetHack, the candles gnomes get are never lit, and the
chance that gnomes get the candles are independent of being the mines or
the lit-ness of the level.