CS877: Add backdoor entry for asmodeus-2
Going along the river instead of crossing the drawbridges will reveal a
wall that is actually a giant mimic, making it possible to skip hoardes
of monsters and directly enter the throne room via a secret passage.

The giant mimic wall looks a bit off, but maybe that's the idea.