CS884: Ask to unlock when opening locked doors
This adds a new option autounlock (on by default), which asks the player
if they want to unlock a locked door when they try to open it and
they're carrying an unlocking tool.

Compared to UnNetHack, this chooses unlocking tools in the order
key/lockpick/card instead of key/card/lockpick to minimize turns spent
unlocking.

This also returns the result of pick_lock() when it's called from
doopen(), so NitroHack can correctly track if an action took place and
therefore a turn should be spent.  This bug may not exist in UnNetHack,
since UnNetHack borrows NetHack's turn logic of relying on flags.move
being set to do the same thing instead.