CS1010: Workaround Vlad's stairs in Wizard's tower
All three Wizard's Tower levels are no flagged noflip to prevent the
stairs to Vlad's Tower appearing within the Wizard's Tower.

The comment for this change refers to "a bug in level flipping", but
doesn't explain exactly what that bug is, nor do the commit messages for
changesets 1010 or 898 (which 1010 supersedes).