Fix wielded potions of blindness conferring resist
Without this fix, wielding a potion of blindness in the main or
alternate weapon slot would grant resistance to certain sources of
blindness, e.g. exploding yellow lights and archon gazes.

Reported with fix by arxanas, confirmed by ais523 and ishanyx:

    http://nethackwiki.com/wiki/Bugs_in_NetHack_3.4.3/Reports#Wielding_potions_of_blindness_confers_blindness_resistance

According to bcode, potions of blindness are the only object in the game
with oc_oprop == BLINDED, strangely enough.