Fix sight-blocking geno'd mimics not unblocking
Without this fix, giant mimics that take forms that block line-of-sight
(e.g. boulders, closed doors) will continue to do so even after being
genocided.

The easiest way to test this is in wizard mode: level teleport to a lit
version of Big Room, create a dozen giant mimics (some will look like
boulders) then wish for and read a scroll of genocide.  The points the
boulder-looking mimics occupied will still block sight to monsters,
which is very obvious with the dark_room patch.

This bug also exists in vanilla NetHack.