Level genocide treats chameleons like normal geno
Normal dungeon-wide genocide treats chameleons as follows:

 1. Genociding the chameleon's true form kills it, no matter what form
    it happens to be at the time.
 2. Genociding the chameleon's current form forces it to change its
    form.

UnNetHack's new behavior for uncursed scrolls of genocide kills all of a
single monster from a level.  However, it treats chameleons differently:

 1. Genociding the chameleon's true form does nothing.
 2. Genociding the chameleon's current form kills it outright.

This commit fixes level genocide to work like dungeon-wide genocide on
chameleons/shape changers, which seems to be what was intended, and
makes it much more useful on e.g. the chameleons on the Rogue home
level.

For reference, mtmp->mnum is the true monster index (i.e. 'chameleon'
for chameleons), monsndx(mtmp->data) get's the monster's current form
(i.e. whatever the chameleon's form is at the time) and mtmp->cham is
set to a non-zero value when a shape changer takes on a form (1 for
chameleon, 2 for doppelganger, 3 for sandestin and 0 when not a shape
changer or a shape changer in their natural form).

This is meant to go with commit af6597c420 (CS122,126,230 nerf scrolls
of genocide).

This bug also exists in UnNetHack.