Fix Rogue level viewed from dungeon overview
Without this fix, choosing to view the Rogue level from the dungeon
overview while not on it will cause it to be displayed without its
'classic' Rogue graphics.  This would also affect future tiles ports if
different tile sets were used for Gehennom, the Gnomish Mines and other
areas.

This is meant to go with commit c461257372 (add remote dungeon level
viewing from the overview menu).

This bug also exists in upstream NitroHack.