Stop autoexplore stopping for unrevealed mimics...
... or monsters with no active attacks, but stop if the monster is
peaceful/tame but can't be known due to hallucination.  These changes
should prevent e.g. object mimics in stores stopping autoexplore, and
allows autoexplore to continue in the presence of the likes of fungi,
floating eyes and gas spores.

The old behavior of stopping autoexplore around hidden mimics is a bug.

This bug also exists in NetHack4.