Fix blind punishment bugs in layered memory system
This adapts the display code for the ball and chain when blind and
punished to NitroHack's new layered memory display, fixing a number of
issues:

 * Object stack inverse video remaining after a felt ball/chain moved
   off a remembered object.

 * Extra object info, e.g. monster info for statues, stack inverse flag,
   not being properly set for an object after a ball/chain was dragged
   over it.

 * Unfelt ball/chain on another object printing the wrong info when
   far-looked, e.g. newt statue with an unfelt chain on it would claim
   to be a giant ant statue instead.

The idea behind the ball and chain code while blind is that the ball and
chain objects override remembered objects on the spaces they occupy, but
keep track of the object they overrided.  When the player knows they've
moved far enough away that the ball/chain has moved, the ball/chain is
removed from the now-obsolete position, and the remembered object memory
of that space is restored.

In NetHack, this system works fine because it remembers whatever glyph
was drawn when the ball/chain moves over a space.  In NitroHack, the
ball/chain overwrites the object, but neglects to save other object info
like monster IDs associated with monster-like objects such as statues.

NitroHack also uses vobj_at() in its textual description of objects from
afar, which returns level->objects[x][y] (INVISIBLE_OBJECTS is not
defined) without considering that an unfelt ball or chain could be at
that position too, which this commit also fixes.

This is meant to go with commit fee7233eb9 (port layered location memory
from SLASH'EM) and commit 644c88b3e7 (CS377,513 multiple items on floor
in inverse video).

This bug also exists in upstream NitroHack.