Say 'blocked' not 'silly' when inv takeoff blocked
Using the inventory object action menu command to take off armor when
it's covered by other armor now says "The rest of your armor is in the
way" instead of "That is a silly thing to take off."

The tricky thing about the new ARMOR_TAKEOFF_BLOCKED value returned by
object_selection_checks() is that it needs the semantics of
ALREADY_IN_USE without triggering the "silly" message in
validate_object().

This is meant to go with commit 240facafb1 (object commands optionally
take argument) and commit 59363d190e (nh_get_object_commands).

This bug also exists in upstream NitroHack.