CS435: pilesize option patch (by Glyn Kennington)
Original patch here: http://bilious.alt.org/?182

Allows setting the number of objects that triggers the "things that are
here" dialog when playing in the classic 80-by-24 interface.  If
pilesize is exceeded, you'll instead get the "there are several objects
here" message.  If set to 1 (the minimum), you'll be told about one
object and never trigger the "things that are here" dialog.

This is not a pure interface change, since setting pilesize low reduces
the chance of feeling cockatrice corpses in object piles.