Color-code turn count changes
The color of the turn count changes based on the number of turns since
the last input:

 * 1 turn - white (as usual)
 * 2-9 turns - green
 * 10-19 turns - blue
 * 20-49 turns - cyan
 * 50-99 turns - yellow
 * 100+ turns - red

The intent is to help players notice the number of turns of certain
actions, e.g. hard engraving, whether a '20s' or '100.' was interrupted
or how long a player gets paralyzed after hitting a floating eye.

This might be too many different colors, but we'll see how this works in
practice.