Fix dragon appearances for polymorph and riding
This fixes the case where wearing the scales of one dragon and
polymorphing into it made you look like another on the map (the game and
description were unaffected), and the case where saddling and riding a
dragon would make it look like another (again, display only).

There's another case in display_self() for M_AP_MONSTER, but I'm not
sure what it's meant to stand for, so I'm not touching it.

This is meant to go with commit 7e66ef6e9d (fix corpse colors for
shuffled dragon appearances).