CS941: Riders/angry quest leaders smash boulders
Strangely, the smashed boulders never leave any rocks behind.

Compared to UnNetHack, the hard-coding of mtmp into is_rockbreaker() is
replaced with a struct monst * versus the struct permonst * it takes in
UnNetHack, and the redundant vision unblocking code for when a monster
picks up a boulder is removed since obj_extract_self() calls
remove_object() for floor objects which does that for boulders already.