CS943: Bag of holding explosion scatters contents
... instead of simply destroying them outright.  Quote from SporkHack
changelog, via UnNetHack commit message:

Be nice to players for a change; just throw their stuff all over the
place if they blow up their bag of holding. ...Of course, it might hit
hard enough to break it. And it might fall in lava. Or water. And it
can't be any fun to try to pick all that stuff back up. :D

According to UnNetHack, this is from SporkHack r525-526.