CS952,953: Fix "dungeon" in drum of earthquake msg
... and use a more appropriate description of the current level instead.

This is what the sky level flag was added for.  In fact, to date, I
believe it's also the only thing it's used for in UnNetHack.

Compared to UnNetHack, get_generic_level_description() takes a struct
level * instead of a struct d_level * since it checks the current
level's sky flag.