CS996: Make minotaurs respect Elbereth
This also makes them vulnerable to scrolls of scare monster.  The
following rationale is from the UnNetHack blog:

http://unnethack.wordpress.com/2012/03/07/why-did-the-minotaur-cross-the-elbereth/

The exception of minotaurs from Elbereth/scare monster was introduced in
NetHack 3.0.5:

https://github.com/bhaak/NetHack/blame/885862f581376d63308c9ccd73420d9b0ae66244/src/monmove.c#LID111

    /* Note: minotaurs must be immune to scare monster to avoid abuse
     * from creating them and taking their wands, then polymorphing 60
     * or so wands to get wishing...
     */

Wands of wishing were excluded as polymorph results in NetHack 3.0.8:

https://github.com/bhaak/NetHack/blame/9a243530575170280fb001b21fbb8ebd18f17499/src/zap.c#LID398

The minotaur comment was removed in NetHack 3.1.0, but its exclusion was
kept all the way up until 3.4.3, the most recent release.