Darken objects/traps/features/walls out of sight
Fixed features like objects, traps and walls are now shown in a darker
shade of their color when out of the field of vision of the character.
I've only ever seen this feature in GruntHack.

This makes use of the visible bit of dbuf, which is a bit bugged when
viewing offscreen levels in the dungeon overview (no fix in this
commit).

This also makes the darkroom option redundant.