Fix unobfuscated object bug in dbuf_set() refactor
obfuscate_object() is meant to work as the final step between level
memory of objects and the display buffer, but I erroneously thought that
obfuscation was done on the memory layer instead, hence the bug, which
is now fixed.

This is meant to go with commit 52fc651bbf (refactor dbuf_set() to
accept -1 for mem defaults).