Autoexplore: Don't favor exploring closed doors
This stops autoexplore from magnetically leading the player into closed
doors, especially locked ones.  This was happening because of the extra
weight that room spaces get during autoexplore also apply to the edges
of the room as well.

This is meant to go with commit 500b497342 (favor room spaces during
autoexplore).