Show weight in inventory object action menu title
If the invweight option is enabled, include the weight of objects in the
title of their action menu, like it appears in the inventory proper.

This also fixes a crash bug in that menu for objects with very long
names (the buffer was too large for the menu system to handle).