Fix call prompt title buffer overflow
This is almost the same as UnNetHack's fix for the same issue with the
naming dialog: the call dialog's title includes the name and type-name
of the object type which could be long enough to overflow the buffer
used to build it.  See also commit 2eddb58f61 (CS660 fix object calling
then naming twice crash).

This bug also exists in upstream NitroHack.