Add name command to inventory object action menu
The name command allows the player to name a single object, or call all
objects of its type if it is permitted.  It appears in the action menu
of all objects except for gold pieces.  (Aside: NitroHack allows gold
pieces to be named, which is a bug.  It doesn't stop stacking, so it's
not too serious.)