Fix extra msg for empty bags when making dumplog
This suppresses the duplicated "sack is empty" message when the dumplog
is created at the end of a game.

This complements commit 00fa98f8ce (hide prompt for empty bags when
creating dumplog).

This is meant to go with commit 28c01f1809 (dumplog, reimplemented).

This bug also exists in upstream NitroHack.