Add unwield to inv action menu of wielded object
"w - unwield (empty)" now appears in the inventory object action menu
for the wielded object.

Internally, this requires support for nh_get_object_commands() to set an
invlet for each object command, the interface to send that invlet back
when setting its next command, and for do_command() to reinterpret '-'
into &zeroobj for command functions that support it (flagged
CMD_ZEROOBJ, currently only engrave and wield).