Color corpses in inventory like on the map
For example, lichen corpses will now appear bright green in the
inventory and other object lists, like they always have when seen on the
map.

This also makes objects' monster numbers available for statues and
figurines in object lists, in case graphical tiles want to draw those
objects in special ways.

This makes a breaking change to the network protocol to send and receive
the new omonnum field of struct nh_objitem.