CS1005: Stethoscopes ID egg type on ground (Spork)
An egg on the ground can now be identified by applying a stethoscope to
it.  This gives a hallucinated monster name if hallucinating, and gives
no info if the egg is stale or empty.  From SporkHack.

Compared to UnNetHack, mons_mname() is used to mask the identity of
randomized dragon eggs.