CS1007-1009: Add Dragon Caves branch (Brass R/S)
The Dragon Caves are a 4-level branch that occurs 10-13 levels deep
within Gehennom.  It consists of predefined cave-like levels with trees,
dead trees, muddy swamps and rivers, and is filled with dragons, worms,
gems, jewelry and gold.

A ring store can be found on its second level, and 3 chromatic dragons
exist at the very last level, which is the only place that non-caveman
characters will find chromatic dragons.

Like Vlad's Tower, monster generation is biased towards chaotic
monsters, and it is considered outside of Gehennom, so it's possible to
e.g. pray safely within it.

According to UnNetHack, this comes from NetHack Brass R/S.

Compared to UnNetHack, the "dragon shop" on the second level is changed
into a "ring shop" (squelching a lev_comp warning and otherwise leaving
an empty room), players get the message that "The heat and smoke are
gone" when entering the branch, and the new Is_dragon(x) macro is added
for future use.