CS1053: hp_notify: show message when HP changes
This shows a message like "[HP-2=10]" when HP changes from one turn to
the next, controlled by the new hp_notify option (default enabled).  The
exact message is controlled by the new hp_notify_format option (default
"[HP%c%a=%h]", using the following format:

 * %a - HP adjustment since last notification, no +/- sign
 * %c - HP adjustment sign, i.e. +/- sign for above
 * %h - current HP
 * %m - maximum HP
 * %H - current HP or "max" if HP is at maximum

Unfortunately the format isn't documented anywhere player-visible.

Compared to UnNetHack, hpnotify_format_str() is renamed to
hp_notify_format_str() with an extra underscore, and contains an extra
line of code to ensure the snprintf() calls never get a negative value.
Also, I'm not 100% sure the notify message is in exactly the same place
as it is in UnNetHack, since NitroHack reworks the main game loop a lot.