CS1056: Make obj menus aware of max object counts
This makes the following improvements to selectable object menus:

 * Selecting with a count adjusts the selection count of a selected
   object instead of deselecting it.
 * Selecting without a count deselects it if it was selected in any way,
   or selects it otherwise.
 * Using a count greater than or equal to the number of objects treats
   it as if it had been selected "normally", i.e. without a count.

Making count-selected objects toggle off instead of on when selected
without a count requires an extra key press to make it fully-selected,
but prevents object groups from becoming part selected/deselected when
selected with then without counts.

Compared to UnNetHack, the implementation is completely different to
cater to NitroHack's Curses UI.  Oddly, the count field of struct
nh_objitem sends object quantities to the front-end, but until now
hasn't taken advantage of it, instead only using it to trasmit player
selection counts back to the game.