CS1078: Shorten Town branch w/ special fill level
The Town branch is now only 2 levels high.  The filler level is a bit
like a big room, with a few random rivers separating the branch exit
from the way up to Town.  It contains 3 random objects, 4 random traps,
5 random monsters and has a 1 in 4 chance of having a giant turtle.

Compared to UnNetHack, townfill is inserted into the dungeon as a LEVEL
versus a RNDLEVEL with 1 file to choose from, and the rocks in the
filler level use quantity, which seems to be what bhaak meant instead of
spe (enchantment).