CS1080: Fix crash w/ exploding bag of holding+gold
Original commit message from UnNetHack r1080:

> Bugfix: Fixed crash when destroying bag of holding containing gold
> pieces
>
> Gold pieces got deallocated erroneously which resulted in a deliberate
> crash as the object's location wasn't OBJ_FREE.

This doesn't crash in NitroHack, since the deliberate crash only occurs
when GOLDOBJ is not defined, which it is in NitroHack.  However,
current_container may not exist any more due to dump_container(), so not
only is delete_contents(current_container) useless, but it could cause a
crash of its own.