CS1081: Prevent dangling pointer w/ destroyed bags
... in doapply().  Actually, doapply() doesn't seem to have any problems
with freed objects, as the case for cream pies in the same switch block
demonstrates, but nothing really important happens after that block, so
there's no harm including this change so long as it's accounted for when
merging NetHack4's fixes to actions taking turns when they shouldn't.

Compared to UnNetHack, this makes the same change for bags of tricks as
well, since putting a wand of cancellation in such a bag (when out of
charges) causes an explosion like with a bag of holding.