Autoexplore: Don't stop for explored object piles
This moves the setting of mem_stepped from domove() to pre_move_tasks(),
allowing code that runs for the current player action to use mem_stepped
to check if the current space has been stepped on before.

pickup() and check_here() now use this info to prevent stopping on
object piles that have been explored before.  In case the object pile
changed since it was last visited, map_object() unsets mem_stepped to
flag it as worth autoexploring again.