Enhance object name detection heuristic
That is, the code that detects if the name given to an object is an
actual name so it can be looked up in data.base when examining them, as
opposed to player note-taking.

To be considered as such, a name now must start with a capital letter,
and also consist of only letters, spaces, "hyphens" (the '-' character)
and single quotes.

Closes #43 on GitHub.