Show inventory weight and used slot count
Both the inventory menu and the sidebar inventory listing now show
"150/900wt (12 of 52 slots)" in the title line, which updates whenever
the player's inventory changes.  The numbers are current weight, maximum
weight before burdened and the number of stacks of objects in the
inventory.

Internally, this info in the sidebar needs updates to player status that
are made for the status lines to occur before drawing the sidebar (read:
curses_update_status() needs to be called before draw_sidebar() /
curses_list_items()), otherwise the info may lag a bit.

Closes #32 on GitHub.