Fix object matching when category menu is skipped
Without this fix, multi-drop will incorrectly say "no matching objects"
with only one object in the inventory when menumatch is set to tight,
and looting a bag with a single object inside says the same thing.

This is meant to go with commit 335ba62d9d (improve matching/filtering
in multi-drop/loot menu).