CS1084: Advent Calendar branch (by Patric Mueller)
Also made available as a patch against vanilla NetHack by the author:

    http://bilious.alt.org/?338

On Dec 1 to Dec 24 inclusive, a portal to the Advent Calendar branch
will appear on dungeon levels 2 to 5.  This takes the player to an open
level with 24 numbered closets, each of which contains a small prize:
candy bar, fortune cookies, rings or tools in decreasing likelihood.

The doors in the Advent Calendar are impassible and indestructible, but
one is unlocked for each day of the Advent Calendar.  The 24th closet
holds a tool that is usually useful (see fill_advent_calendar() for
details).

Compared to UnNetHack:

 * advcal-2 is removed, since it's unused, and doesn't have enough doors
   or closets to adapt to work with the Advent Calendar code.
 * "re-enforced" -> "reinforced"